update op unit tests
