+++ tests/test_tensor/test_addmm_tp.py
@@ -11,6 +11,7 @@ from colossalai.testing import rerun_if_address_is_in_use
 from colossalai.utils import free_port
 from functools import partial
 from colossalai.core import global_context as gpc
+from _utils import tensor_shard_equal, tensor_equal
 
 
 class Conv1D(nn.Module):
@@ -45,13 +46,6 @@ def init_1d_row(weight, bias):
         weight.set_spec(spec)
 
 
-def check_grad_1d_row(model: torch.nn.Module, weight, bias):
-    rank = gpc.get_local_rank(ParallelMode.PARALLEL_1D)
-    size = gpc.get_world_size(ParallelMode.PARALLEL_1D)
-    assert torch.allclose(model.weight.grad.chunk(size, 0)[rank], weight.grad)
-    assert torch.allclose(model.bias.grad, bias.grad)
-
-
 def init_1d_col(weight, bias):
     spec = TensorSpec(
         distspec.shard(gpc.get_group(ParallelMode.PARALLEL_1D), [-1], [gpc.get_world_size(ParallelMode.PARALLEL_1D)]),
@@ -61,14 +55,7 @@ def init_1d_col(weight, bias):
         bias.set_spec(spec)
 
 
-def check_grad_1d_col(model: torch.nn.Module, weight, bias):
-    rank = gpc.get_local_rank(ParallelMode.PARALLEL_1D)
-    size = gpc.get_world_size(ParallelMode.PARALLEL_1D)
-    assert torch.allclose(model.weight.grad.chunk(size, -1)[rank], weight.grad)
-    assert torch.allclose(model.bias.grad.chunk(size, -1)[rank], bias.grad)
-
-
-def run_with_spec(spec_init_func, check_grad_func):
+def run_with_spec(spec_init_func):
     model = Conv1D(4, 16).cuda()
     weight = ColoTensor(torch.nn.Parameter(model.weight.detach()))
     bias = ColoTensor(torch.nn.Parameter(model.bias.detach()))
@@ -76,18 +63,19 @@ def run_with_spec(spec_init_func, check_grad_func):
     x = torch.rand(2, 16).cuda()
     out = model(x)
     colo_out = torch.addmm(bias, x, weight)
-    assert torch.allclose(out, colo_out)
+    assert tensor_equal(out, colo_out)
     grad = torch.rand_like(out)
     out.backward(grad)
     colo_out.backward(grad)
-    check_grad_func(model, weight, bias)
+    tensor_shard_equal(model.weight.grad, weight.grad)
+    tensor_shard_equal(model.bias.grad, bias.grad)
 
 
 def run_dist(rank, world_size, port):
     config = dict(parallel=dict(tensor=dict(mode="1d", size=world_size),))
     colossalai.launch(config=config, rank=rank, world_size=world_size, host='localhost', port=port, backend='nccl')
-    run_with_spec(init_1d_row, check_grad_1d_row)
-    run_with_spec(init_1d_col, check_grad_1d_col)
+    run_with_spec(init_1d_row)
+    run_with_spec(init_1d_col)
 
 
 @pytest.mark.dist
