+++ tests/test_tensor/test_embedding_tp.py
@@ -12,6 +12,7 @@ from colossalai.testing import rerun_if_address_is_in_use
 from colossalai.utils import free_port
 from colossalai.core import global_context as gpc
 from colossalai.tensor import TensorSpec, ComputePattern, ParallelAction, DistSpecManager
+from _utils import tensor_equal, tensor_shard_equal
 
 
 def init_1d_row(weight):
@@ -22,12 +23,6 @@ def init_1d_row(weight):
         weight.set_spec(spec)
 
 
-def check_grad_1d_row(model: torch.nn.Module, weight):
-    rank = gpc.get_local_rank(ParallelMode.PARALLEL_1D)
-    size = gpc.get_world_size(ParallelMode.PARALLEL_1D)
-    assert torch.allclose(model.weight.grad.chunk(size, 0)[rank], weight.grad)
-
-
 def init_1d_col(weight):
     spec = TensorSpec(
         distspec.shard(gpc.get_group(ParallelMode.PARALLEL_1D), [-1], [gpc.get_world_size(ParallelMode.PARALLEL_1D)]),
@@ -36,31 +31,25 @@ def init_1d_col(weight):
         weight.set_spec(spec)
 
 
-def check_grad_1d_col(model: torch.nn.Module, weight):
-    rank = gpc.get_local_rank(ParallelMode.PARALLEL_1D)
-    size = gpc.get_world_size(ParallelMode.PARALLEL_1D)
-    assert torch.allclose(model.weight.grad.chunk(size, -1)[rank], weight.grad)
-
-
-def run_with_spec(spec_init_func, check_grad_func):
+def run_with_spec(spec_init_func):
     model = torch.nn.Embedding(12, 32).cuda()
     weight = ColoTensor(torch.nn.Parameter(model.weight.detach()))
     spec_init_func(weight)
     x = torch.tensor((0, 3, 6, 9)).cuda()
     out = model(x)
     colo_out = F.embedding(x, weight)
-    assert torch.allclose(out, colo_out)
+    assert tensor_equal(out, colo_out)
     grad = torch.rand_like(out)
     out.backward(grad)
     colo_out.backward(grad)
-    check_grad_func(model, weight)
+    assert tensor_shard_equal(model.weight.grad, weight.grad)
 
 
 def run_dist(rank, world_size, port):
     config = dict(parallel=dict(tensor=dict(mode="1d", size=world_size),))
     colossalai.launch(config=config, rank=rank, world_size=world_size, host='localhost', port=port, backend='nccl')
-    run_with_spec(init_1d_row, check_grad_1d_row)
-    run_with_spec(init_1d_col, check_grad_1d_col)
+    run_with_spec(init_1d_row)
+    run_with_spec(init_1d_col)
 
 
 @pytest.mark.dist
